refactor(download): 拆分useDownload为常量+初始化+操作三层
